+++ doubleml/iv/__init__.py
@@ -2,7 +2,7 @@
 The :mod:`doubleml.iv` module implements double machine learning estimates based on instrumental variable models.
 """
 
-from .pliv import DoubleMLPLIV
+from ..plm.pliv import DoubleMLPLIV
 from .iivm import DoubleMLIIVM
 
 __all__ = [
