+++ doubleml/plm/__init__.py
@@ -1,0 +1,11 @@
+"""
+The :mod:`doubleml.plm` module implements double machine learning estimates based on partially linear models.
+"""
+
+from .plr import DoubleMLPLR
+from .pliv import DoubleMLPLIV
+
+__all__ = [
+    "DoubleMLPLR",
+    "DoubleMLPLIV",
+]
